OMAP4: SMC: Fix compiler warnings in omap4-common.c
arch/arm/mach-omap2/omap4-common.c:40: warning: 'l4_secure_clkdm'
	defined but not used

If SMC is defined, we use a different omap4_secure_dispatcher and
l4_secure_clkdm was being defined but not used in omap4-common.c. The
solution here is to wrap the declaration in same #ifdef as
omap4_secure_dispatcher.

Change-Id: Idd0a7c144c78f30d2ea7269390cad9112e218659
Signed-off-by: Bryan Buckley <bryan.buckley@ti.com>