cpufreq: hotplug: don't synch threads on jiffies
cpufreq-hotplug was always designed to run as a single thread that takes
aggregate load of all CPUs into account.  As such there is no need to
synchronize multiple threads to execute at the same time.  This
sychronization mechanism also causes execution of the do_dbs_check function
to be somewhat erratically scheduled.  Removing it causes scheduling of this
function to reliably happen based soley on the supplied sampling_rate
parameter.

Change-Id: If0edc89996f0d5763502236b225835dd9f4ab3c9
Signed-off-by: Mike Turquette <mturquette@ti.com>